gameplay corrections

+++ src/Headers/MapCollision.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-bool map_collision(bool i_collect_pellets, bool i_use_door, short i_x, short i_y, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WIDTH>& i_map);
+bool map_collision(bool i_collect_pellets, bool i_use_door, int i_x, int i_y, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WIDTH>& i_map);
